+++ RedEye/Engine/RE_CompMesh.h
@@ -16,7 +16,7 @@ public:
 	const char* GetMesh() const;
 
 	unsigned int GetVAOMesh()const;
-	unsigned int GetTriangleMesh()const;
+	size_t GetTriangleMesh()const;
 
 	void SetMaterial(const char* md5);
 	const char* GetMaterial()const;
@@ -26,7 +26,7 @@ public:
 	void SerializeJson(RE_Json* node, eastl::map<const char*, int>* resources) const override;
 	void DeserializeJson(RE_Json* node, eastl::map<int, const char*>* resources) override;
 
-	unsigned int GetBinarySize()const override;
+	size_t GetBinarySize() const override;
 	void SerializeBinary(char*& cursor, eastl::map<const char*, int>* resources) const override;
 	void DeserializeBinary(char*& cursor, eastl::map<int, const char*>* resources) override;
